+++ src/main/java/com/snowflake/snowpark_java/stub/Broadcast.java
@@ -1,0 +1,18 @@
+package com.snowflake.snowpark_java.stub;
+
+@Deprecated
+public class Broadcast<T> {
+
+    public T value() {
+        return null;
+    }
+
+    public T getValue() {
+        return null;
+    }
+
+    public static <T> Broadcast<T> ofValue(T value) {
+        throw new RuntimeException("Not implemented");
+    }
+    
+}