+++ src/main/java/com/snowflake/snowpark_java/stub/MapFunction.java
@@ -1,0 +1,8 @@
+package com.snowflake.snowpark_java.stub;
+
+@Deprecated
+@FunctionalInterface
+public interface MapFunction<T, U>  {
+  U call(T value) throws Exception;
+}
+