+++ src/main/java/com/snowflake/snowpark_java/stub/MapGroupsFunction.java
@@ -1,0 +1,12 @@
+package com.snowflake.snowpark_java.stub;
+
+import java.util.Iterator;
+
+/**
+ * Base interface for a map function used in GroupedDataset's mapGroup function.
+ */
+@Deprecated
+@FunctionalInterface
+public interface MapGroupsFunction<K, V, R>  {
+  R call(K key, Iterator<V> values) throws Exception;
+}